stream-engine: extract public definitions
