+++ src/stream-engine/common/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "stream-engine-common"
+version = "0.1.0"
+edition = "2021"
+publish = false
+
+[dependencies]
+thiserror = "1.0"
+tonic = "0.6"
