+++ src/stream-engine/master/Cargo.toml
@@ -5,6 +5,7 @@ edition = "2021"
 publish = false
 
 [dependencies]
+stream-engine-common = { version = "0.1", path = "../common" }
 stream-engine-proto = { version = "0.1", path = "../proto" }
 
 log = "0.4.14"
