+++ src/stream-engine/store/Cargo.toml
@@ -5,6 +5,7 @@ edition = "2021"
 publish = false
 
 [dependencies]
+stream-engine-common = { version = "0.1", path = "../common" }
 stream-engine-proto = { version = "0.1", path = "../proto" }
 
 futures = "0.3.21"
